+++ src/main/java/com/example/springrider/domain/order/dto/request/CreateOrderRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.springrider.domain.order.dto;
+package com.example.springrider.domain.order.dto.request;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
