+++ src/main/java/com/example/springrider/domain/order/dto/response/OrderResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.springrider.domain.order.dto;
+package com.example.springrider.domain.order.dto.response;
 
 import com.example.springrider.domain.order.entity.Order;
 import java.util.List;
