+++ src/main/java/com/example/springrider/domain/order/service/UserOrderService.java
@@ -5,8 +5,8 @@ import com.example.springrider.domain.cart.entity.CartItem;
 import com.example.springrider.domain.cart.repository.CartRepository;
 import com.example.springrider.domain.common.exception.ExceptionCode;
 import com.example.springrider.domain.common.exception.InvalidRequestException;
-import com.example.springrider.domain.order.dto.CreateOrderRequestDto;
-import com.example.springrider.domain.order.dto.OrderResponseDto;
+import com.example.springrider.domain.order.dto.request.CreateOrderRequestDto;
+import com.example.springrider.domain.order.dto.response.OrderResponseDto;
 import com.example.springrider.domain.order.entity.Order;
 import com.example.springrider.domain.order.entity.OrderItem;
 import com.example.springrider.domain.order.enums.OrderStatus;
